fix tests 4

+++ tests/libtest/cli_ws_data.c
@@ -106,7 +106,7 @@ static CURLcode test_ws_data_m2_echo(const char *url,
   curl_easy_setopt(curl, CURLOPT_VERBOSE, 1L);
   curl_easy_setopt(curl, CURLOPT_CONNECT_ONLY, 2L); /* websocket style */
   r = curl_easy_perform(curl);
-  curl_mfprintf(stderr, "curl_easy_perform() returned %u\n", (int)r);
+  curl_mfprintf(stderr, "curl_easy_perform() returned %u\n", r);
   if(r != CURLE_OK)
     goto out;
